+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Bug Fixes
 
+- Fix breadcrumb sanitization: emit bugsnag_unserializable_values as string[], preserve warnings.
+[#975](https://github.com/bugsnag/bugsnag-unity/pull/975)
+
 - Fix Android context being overwritten by Activity
 [#977](https://github.com/bugsnag/bugsnag-unity/pull/977)